moving trace classes
