+++ src/jdom/jacdac-jdom.ts
@@ -25,12 +25,12 @@ export * from "./clients/sensoraggregatorclient"
 export * from "./iframeclient"
 export * from "./iframebridgeclient"
 export * from "./frame"
-export * from "./trace"
-export * from "./traceplayer"
+export * from "./trace/trace"
+export * from "./trace/traceplayer"
 export * from "./command"
 export * from "./packetfilter"
-export * from "./tracerecorder"
-export * from "./traceview"
+export * from "./trace/tracerecorder"
+export * from "./trace/traceview"
 export * from "./flags"
 export * from "./pack"
 export * from "./packobject"
