+++ typedoc.json
@@ -15,7 +15,7 @@
     "excludeInternal": true,
     "excludeExternals": true,
     "name": "Jacdac TypeScript",
-    "categoryOrder": ["JDOM", "Clients", "*"],
+    "categoryOrder": ["JDOM", "Clients", "Servers", "Trace", "*"],
     "readme": "none",
     "includeVersion": true
 }
